CLN: move get_regularization from util to sr3
